+++ components/ota_upload/include/ota_upload.h
@@ -12,12 +12,16 @@
 extern "C" {
 #endif
 
+typedef esp_err_t (*ota_upload_prepare_cb_t)(void *ctx);
+
 typedef struct
 {
   const char *hostname;
   uint16_t port;
   bool enable_mdns;
   bool reboot_after_update;
+  ota_upload_prepare_cb_t prepare_cb;
+  void *prepare_ctx;
 } ota_upload_config_t;
 
 #define OTA_UPLOAD_CONFIG_DEFAULT()                    \
@@ -25,7 +29,9 @@ typedef struct
     .hostname = "esp-ota",                              \
     .port = CONFIG_OTA_UPLOAD_PORT,                    \
     .enable_mdns = CONFIG_OTA_UPLOAD_ENABLE_MDNS,      \
-    .reboot_after_update = CONFIG_OTA_UPLOAD_REBOOT_AFTER_UPDATE \
+    .reboot_after_update = CONFIG_OTA_UPLOAD_REBOOT_AFTER_UPDATE, \
+    .prepare_cb = NULL,                                  \
+    .prepare_ctx = NULL                                  \
   }
 
 esp_err_t ota_upload_start(const ota_upload_config_t *config);