+++ host_tools/src/ota_upload_tools/__init__.py
@@ -1,4 +1,4 @@
 #-- SPDX-License-Identifier: GPL-3.0-or-later
 #-- Copyright (C) 2026 Willem Aandewiel
 
-__version__ = "0.1.0"
+__version__ = "1.0.0"